Code restructurization, folders reorganization
- All the generators code was moved to other file from views.py to keep "requests handlers" and "generators" separated, because 200 lines of code in one view is not so good;
- _temporary_images/ folder is now auto-clearing on runserver exit in DEBUG=True;
- unnecessary files deleted.
